Use autogenerated client in datatypes service
