+++ client/src/components/Datatypes/index.js
@@ -1,1 +1,1 @@
-export { getDatatypesMapper } from "./factory.js";
+export { getDatatypesMapper } from "./factory";
