+++ client/src/components/Datatypes/index.test.js
@@ -1,23 +1,13 @@
-import axios from "axios";
-import MockAdapter from "axios-mock-adapter";
-
 import { getDatatypesMapper } from "./index";
+import { getDatatypes } from "./services";
 import { typesAndMappingResponse } from "./test_fixtures";
 
+jest.mock("./services");
+
 describe("Datatypes/index.js", () => {
-    let axiosMock;
-
-    beforeEach(() => {
-        axiosMock = new MockAdapter(axios);
-    });
-
-    afterEach(() => {
-        axiosMock.restore();
-    });
-
     describe("getDatatypesMapper", () => {
         it("should fetch logic from API for comparing datatypes in a hierarchy", async () => {
-            axiosMock.onGet(`/api/datatypes/types_and_mapping`).reply(200, typesAndMappingResponse);
+            getDatatypes.mockResolvedValue(typesAndMappingResponse);
             await getDatatypesMapper().then((mapper) => {
                 expect(mapper.isSubType("txt", "data")).toBe(true);
                 expect(mapper.isSubType("txt", "txt")).toBe(true);
