+++ client/src/components/Datatypes/test-data/datatypes.json
@@ -1,0 +1,1 @@
+../../../../tests/qunit/test-data/json/datatypes.json
