+++ client/src/components/Datatypes/test-data/datatypes.mapping.json
@@ -1,0 +1,1 @@
+../../../../tests/qunit/test-data/json/datatypes.mapping.json
